refactor: outsource types
