+++ apps/dsp-app/src/app/project/ontology/ontology-form/ontology-form.type.ts
@@ -1,0 +1,12 @@
+import { FormControl, FormGroup } from '@angular/forms';
+
+export interface OntologyFormProps {
+  projectIri: string;
+  ontologyIri?: string;
+}
+
+export type OntologyForm = FormGroup<{
+  name: FormControl<string>;
+  label: FormControl<string>;
+  comment: FormControl<string>;
+}>;
